+++ client/src/Components/index.jsx
@@ -1,11 +1,3 @@
-import AddLocationModal from './AddLocationModal';
-import EditLocationModal from './EditLocationModal';
-import AddDeviceModal from './AddInventoryModal';
-import EditDeviceModal from './EditInventoryModal';
-import AddDonorModal from './AddDonorModal';
-import EditDonorModal from './EditDonorModal';
 import DeleteModal from './DeleteModal';
-import AddCourseModal from './AddCourseModal';
-import EditCourseModal from './EditCourseModal';
 
-export { AddLocationModal, EditLocationModal, DeleteModal, AddDeviceModal, EditDeviceModal, AddCourseModal, EditCourseModal, AddDonorModal, EditDonorModal};
+export { DeleteModal };
