--- client/src/Pages/index.jsx
@@ -1,9 +1,0 @@
-import Appointments from './Appointments';
-import Tickets from './Tickets';
-import Clients from './Clients';
-import Inventory from './Inventory';
-import Locations from './Locations';
-import Donors from './Donors';
-import Courses from './Courses';
-
-export { Appointments, Tickets, Clients, Inventory, Locations, Donors, Courses };
